fix: UserRepository Interface 와 구현체로 변경 (for test)
